+++ packages/render-composer/src/index.ts
@@ -1,12 +1,8 @@
+export * from "./passes"
 export * from "./Canvas"
-export * from "./CopyPass"
-export * from "./DepthCopyPass"
 export * from "./EffectComposer"
-export * from "./EffectPass"
 export * from "./GodRaysEffect"
-export * from "./LayerRenderPass"
 export * from "./Layers"
-export * from "./RenderPass"
 export * from "./RenderPipeline"
 export * from "./SelectiveBloomEffect"
 export * from "./SMAAEffect"
