+++ packages/render-composer/src/passes/CopyPass.tsx
@@ -1,6 +1,6 @@
 import * as PP from "postprocessing"
 import { forwardRef, useContext, useImperativeHandle, useMemo } from "react"
-import { EffectComposerContext } from "./EffectComposer"
+import { EffectComposerContext } from "../EffectComposer"
 
 export const CopyPass = forwardRef<PP.CopyPass>((_, ref) => {
   const pass = useMemo(() => new PP.CopyPass(), [])
