+++ packages/render-composer/src/passes/DepthCopyPass.tsx
@@ -1,7 +1,7 @@
 import * as PP from "postprocessing"
 import { forwardRef, useContext, useImperativeHandle, useMemo } from "react"
 import { BasicDepthPacking } from "three"
-import { EffectComposerContext } from "./EffectComposer"
+import { EffectComposerContext } from "../EffectComposer"
 
 export const DepthCopyPass = forwardRef<PP.DepthCopyPass>((_, ref) => {
   const pass = useMemo(
