+++ packages/render-composer/src/passes/RenderPass.tsx
@@ -1,7 +1,7 @@
 import { useThree } from "@react-three/fiber"
 import { RenderPass as RenderPassImpl } from "postprocessing"
 import { useContext, useMemo } from "react"
-import { EffectComposerContext } from "./EffectComposer"
+import { EffectComposerContext } from "../EffectComposer"
 
 export const RenderPass = () => {
   const scene = useThree((s) => s.scene)
