+++ packages/render-composer/src/passes/index.ts
@@ -1,0 +1,5 @@
+export * from "./CopyPass"
+export * from "./DepthCopyPass"
+export * from "./EffectPass"
+export * from "./LayerRenderPass"
+export * from "./RenderPass"
